Data_Set_used
This was the data set used with the stored procedures. The randomness of the data does not reflect a true representation. HESid had varying numbers of episodes but listed as one spell so made the assumption to group together the episodes by HESID and disregard the spell column. Random HRG codes have been assigned as well so strange values will be obtained.